requirements.txt
asgiref==3.9.1
beautifulsoup4==4.13.5
Brotli==1.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
dj-database-url==3.0.1
Django==5.2.5
django-cors-headers==4.8.0
djangorestframework==3.16.1
gunicorn==23.0.0
h11==0.16.0
idna==3.10
packaging==25.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
requests==2.32.5
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
whitenoise==6.10.0
